Move Table component to DataViz dir
