+++ src/components/DataViz/index.js
@@ -1,2 +1,4 @@
 export { PieChart } from './PieChart'
 export { BarChart } from './BarChart'
+export { Table } from './Table'
+export { TableActionButtons, TablePagingButtons } from './TableButtons'
